tegra: camera: rtcpu: update firmware api headers
tag 57be429b313e2bab6389a93ab22d4d2a3c12a2b5

rtcpu@f3fc3a6
    8ea9f2e rtcpu: Misc MISRA fixes
    4f9cf92 rtcpu: rtos_result in tegra-vi & related MISRA fixes

Bug 2016681
Bug 2060930

Reviewed-on: https://git-master.nvidia.com/r/1667079
GVS: Gerrit_Virtual_Submit
Tested-by: Pekka Pessi <ppessi@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ include/soc/tegra/camrtc-common.h
@@ -27,10 +27,10 @@
 #define U32_C(x) x##UL
 #endif
 #ifndef U16_C
-#define U16_C(x) x##U
+#define U16_C(x) (uint16_t)(x##U)
 #endif
 #ifndef U8_C
-#define U8_C(x) x##U
+#define U8_C(x) (uint8_t)(x##U)
 #endif
 #endif
